Refactor function to multiple smaller ones
